requirements.txt
Pillow==5.0.0
pymssql==2.1.3
pytesseract==0.2.0
selenium==3.9.0
